File renaming and header inclusion cleanup; NFC
