+++ src/AST.cpp
@@ -1,7 +1,7 @@
-#include "ast.h"
-#include "llvm/Support/raw_ostream.h"
+#include "AST.h"
 
 #include <llvm/Support/Casting.h>
+#include <llvm/Support/raw_ostream.h>
 
 using namespace ast;
