+++ src/Runtime.cpp
@@ -1,4 +1,4 @@
-#include "runtime.h"
+#include "Runtime.h"
 
 #include <llvm/Support/Casting.h>
