+++ src/include/ASTVisitor.h
@@ -2,7 +2,7 @@
 
 #include <memory>
 
-#include "ast_fwd.h"
+#include "AST_fwd.h"
 
 class ASTVisitor {
 public:
