+++ test/unit/CMakeLists.txt
@@ -19,13 +19,13 @@ llvm_map_components_to_libnames(LLVM_LIBS support)
 
 add_executable(test_parse
   test_parse.cpp
+  ${PROJECT_SOURCE_DIR}/src/AST.cpp
+  ${PROJECT_SOURCE_DIR}/src/IdPool.cpp
   ${PROJECT_SOURCE_DIR}/src/Lex.cpp
   ${PROJECT_SOURCE_DIR}/src/Parse.cpp
-  ${PROJECT_SOURCE_DIR}/src/UTF8.cpp
+  ${PROJECT_SOURCE_DIR}/src/Runtime.cpp
   ${PROJECT_SOURCE_DIR}/src/SourceStream.cpp
-  ${PROJECT_SOURCE_DIR}/src/ast.cpp
-  ${PROJECT_SOURCE_DIR}/src/idpool.cpp
-  ${PROJECT_SOURCE_DIR}/src/runtime.cpp
+  ${PROJECT_SOURCE_DIR}/src/UTF8.cpp
 )
 target_link_libraries(test_parse PRIVATE Catch2::Catch2 gmp gmpxx ${LLVM_LIBS})
